requirements.txt
crewai 
unstructured ==0.10.25
pyowm
python-dotenv
streamlit
langchain
request
pydantic 
